trendsync-backend/requirements.txt
# Core AI SDK - manages all web framework and AI dependencies
google-adk==1.18.0

# Google Cloud
google-cloud-storage>=2.14.0

# Utilities
python-dotenv==1.0.1
httpx>=0.27.0
python-docx>=1.1.0
redis>=5.0.0
Pillow>=10.0.0
requests>=2.31.0
websockets>=12.0
